Add some tests for variations
