+++ setup.cfg
@@ -1,3 +1,5 @@
 [flake8]
 exclude = .git,build,deap.egg-info,doc
 max-line-length = 88
+per-file-ignores =
+    deap/tests/*:D100,D101,D102
